// app/products/webapp/utils/enums/ProductStatus.ts
export enum ProductStatus {
    InProgress = "InProgress",
    NewSupplierProduct = "NewSupplierProduct",
    ToCheck = "ToCheck",
    RequestedToSAP = "RequestedToSAP",
    CreationFailed = "CreationFailed",
    CreatedInSAP = "CreatedInSAP",
    MarkedForDeletion = "MarkedForDeletion",
    ReleasedForSupplier = "ReleasedForSupplier",
    PartiallyCreatedInSAP = "PartiallyCreatedInSAP"
}

export enum ArticleStatus {
    InProgress = "InProgress",
    NewSupplierProduct = "NewSupplierProduct",
    ToCheck = "ToCheck",
    RequestedToSAP = "RequestedToSAP",
    CreationFailed = "CreationFailed",
    CreatedInSAP = "CreatedInSAP",
    MarkedForDeletion = "MarkedForDeletion",
    ReleasedForSupplier = "ReleasedForSupplier",
    PartiallyCreatedInSAP = "PartiallyCreatedInSAP"
}

export enum VariantStatus {
    InProgress = "InProgress",
    NewSupplierProduct = "NewSupplierProduct",
    ToCheck = "ToCheck",
    RequestedToSAP = "RequestedToSAP",
    CreationFailed = "CreationFailed",
    CreatedInSAP = "CreatedInSAP",
    MarkedForDeletion = "MarkedForDeletion",
    ReleasedForSupplier = "ReleasedForSupplier",
    PartiallyCreatedInSAP = "PartiallyCreatedInSAP"
}